move demo plugins into own directory
